refactor: move rich training loggers to unilab.logging
