+++ src/unilab/algos/torch/offpolicy/__init__.py
@@ -3,7 +3,7 @@
 from unilab.algos.torch.offpolicy.multi_gpu_runner import MultiGPUOffPolicyRunner
 from unilab.algos.torch.offpolicy.runner import OffPolicyRunner
 from unilab.algos.torch.offpolicy.worker import off_policy_collector_fn
-from unilab.training.logging.offpolicy import OffPolicyLogger
+from unilab.logging import OffPolicyLogger
 
 __all__ = [
     "OffPolicyLogger",
