+++ tests/utils/test_experiment_tracking.py
@@ -4,9 +4,8 @@ import json
 import sys
 from pathlib import Path
 
-from unilab.training.logging.experiment import ExperimentTracker, build_wandb_settings
-from unilab.training.logging.offpolicy import OffPolicyLogger
-from unilab.training.logging.onpolicy import OnPolicyLogger
+from unilab.logging import OffPolicyLogger, OnPolicyLogger
+from unilab.training.experiment import ExperimentTracker, build_wandb_settings
 
 
 class _FakeConfig(dict):
